Apply .gitignore

+++ mixer.h
@@ -2,4 +2,4 @@
 #include <gz/msgs/actuators.pb.h>
 #include <gz/transport/Node.hh>
 
-float mixer(gz::transport::Node::Publisher &motorPub, double output[4]);
+float mixer(gz::transport::Node::Publisher &motorPub, float output[4]);
